requirements.txt
requests>=2.28.0
chromadb>=0.4.0
onnxruntime>=1.16.0
tokenizers>=0.15.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
google-genai>=1.0.0
